requirements.txt
# =============================================================================
# DÉPENDANCES NARRANDO
# =============================================================================

# Dépendances de base
requests==2.32.4
python-dotenv==1.1.1
argparse==1.4.0
openai==1.40.6
sentry-sdk==2.14.0
pytest==8.3.3

# API et base de données
supabase==2.7.4
flask==3.0.0
flask-cors==4.0.0

# Traitement de données
polyline==2.0.0

# AWS Services
boto3==1.34.162

# Serveur WSGI (VPS / containers)
gunicorn==21.2.0

# Audio Generation
elevenlabs==2.14.0

# Utilitaires (modules intégrés Python)
# datetime, json, math, random, os, time, typing, subprocess sont des modules intégrés
